src: Separate SubReadStream and SeekableSubReadStream into different files.
