+++ src/libraries/System.IO.Compression/src/System.IO.Compression.csproj
@@ -29,7 +29,6 @@
              Link="Common\System\IO\Compression\ZLibNative.cs" />
     <Compile Include="$(CommonPath)System\IO\Compression\ZLibNative.ZStream.cs"
              Link="Common\System\IO\Compression\ZLibNative.ZStream.cs" />
-    <Compile Include="System\IO\Compression\CompressionCustomStreams.cs" />
     <Compile Include="System\IO\Compression\CompressionLevel.cs" />
     <Compile Include="System\IO\Compression\CompressionMode.cs" />
     <Compile Include="System\IO\Compression\Crc32Helper.ZLib.cs" />
@@ -46,6 +45,8 @@
     <Compile Include="System\IO\Compression\TarArchiveMode.cs" />
     <Compile Include="System\IO\Compression\TarArchiveOptions.cs" />
     <Compile Include="System\IO\Compression\TarFile.cs" />
+    <Compile Include="System\IO\Compression\SubReadStream.cs" />
+    <Compile Include="System\IO\Compression\SeekableSubReadStream.cs" />
     <Compile Include="System\IO\Compression\ZLibStream.cs" />
     <Compile Include="$(CommonPath)System\Threading\Tasks\TaskToApm.cs"
              Link="Common\System\Threading\Tasks\TaskToApm.cs" />
